Moved a lot of logic around to match what I deduce is (closer to) how and when the real UITableView loads its own internal caches. This is all rather "clever" but I think it works. The best way to know for sure is to get it into the big project so Ostrich can find any edge cases.
The table view should now allow you to select a row even before it was reloaded as long as there's a dataSource (which the real one does as well), and it should also still work right after a reload but before any cells were created and or the table was made visible. Very tricky stuff here.

Also implemented proper scroll position support.

UITableView is now diamonds.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4258 d3437ac4-8b1b-0410-be28-b75ab53cb89f