Fixed a very subtle crash. This could happen if you clicked the "load next reply" button in a conversation view. The reason for the crash is complex.
The "load next reply" button ultimately gets released while its touchesEnded is being handled because the entire button itself is removed from the table view and replaced with a status message view. When the button is removed, it's removeFromSuperview implementation (in UIView) ends up calling the _cancelTouchesInView: method. When it does so, since it was the touch's active view that is being removed from the superview, it canceled the touch and set the touch's view to nil - thus releasing itself from the touch object. That release was the final release necessary to destroy the button object itself. The button then dealloced. Eventually control returns to UIControl which tries to send itself a message as part of the final touch ended code. This crashes and burns because "self" (the original "load next reply" button) has already been destroyed.

Crash bang boom.

The only reason the touch's view was being set to nil here was to get the mouse enter/exit stuff to work in a particular way - namely when touches were cancelled, it'd act as if the mouse moved out of the window entirely and waited for the user to move the mouse (or click or scroll or whatever) before the enter/exit stuff started working again. Now, to avoid this obscure crash, it will not do this. Instead it relies on the next mouse event to trigger the enter/exit test logic without any intervening fake cursor exiting events. This has the side effect of sometimes causing a widget or what-have-you to linger longer than might seem fitting after a menu is dismissed or a popover is displayed. The cursor needs to move into a UIKit view before the enter/exit events get triggered again. This isn't the end of the world, though, and I'm not even certain if one way or the other is really "right." With the previous approach, even if the mouse cursor ended up hovering over the same view it was in before the touch was cancelled, you'd have to move the cursor again to get the enter/exit stuff to trigger anyway. So really, either the oddness is at the start of a new mouse move, or the end of the previous one. I'm not sure it matters which it is.

Anyway, this should avoid this one particular crash, but who knows if it will end up revealing a different sort of problem later on...

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4296 d3437ac4-8b1b-0410-be28-b75ab53cb89f