It now uses an edge inset instead of having to override textRectForBounds: to adjust margins which removes a potential need to subclass just to get some custom layout behavior. This also allows for sizeToFit to work and take the edge insets into account when returning it's new size.
I also split out the core text framesetter and frame a bit so that it doesn't recreate everything from scratch as often. The big win here is that when the core text's frame is resized, it no longer does nearly as much work because the typesetter isn't destroyed and recreated constantly. It also allows any calls to sizeToFit to reuse the existing framesetter (if any) so that wastes less time, too.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4415 d3437ac4-8b1b-0410-be28-b75ab53cb89f