This is a huge and complex change which adds support for keyboard events in general and specifically allowed for a way to handle dismissing and sending the compose window in Ostrich.
I have defined a new UIResponder method named:
- (void)keyPressed:(UIKey *)key withEvent:(UIEvent *)event;

There's also a new UIKit class named UIKey which isn't based on anything in the actual UIKit (as far as I'm aware) and so breaks with Apple's design here since Apple has yet to publish any keyboard event handling.

In addition to handling keyboard events and passing them down the responder chain and attempting to define them in a relatively clean way, I also defined two "standard" responder messages similar to cut:/copy:/paste: in that they are only specified as an informal protocol, but the system will map two specific keyboard events to them. They are named: cancel:/commit:.

cancel: is sent whenever AppKit would normally send canceOperation: This means that pressing ESC or CMD-. will generally result in an attempt at sending cancel: to the current first responder.

commit: is sent whenever my UIKit captures CMD-Return or Enter.

These messages work just like copy:/paste:/etc. First, canPerformAction:withSender: is used on the key window's first responder to determine if either message should even be sent. If the answer to that is YES, it is then sent to the first object along the responder chain (starting at the key window's first responder) that can handle it. Otherwise the messages are not sent. The upshot of this is that UIKit views or view controllers need only define methods named cancel: or commit:, and the right thing should generally happen.

There is significant complexity in routing these messages around from within the UITextView (which is really just an NSTextView that's all wrapped up with a bow). This has complicated an already complicated situation here.

As an added bonus, I've also fixed (I think) some other responder-related problems. Specifically, clicking in a UITextView to make it the first responder would not cause the UITextView itself to become the first responder from the point of view of UIKit (instead only the NSTextView that is wrapped was the first responder.. which meant there was potential for UIKit-based code to be very confused about some things). Things should now be working such that no matter how a UITextView (or field) become the first responder, both AppKit and UIKit should be agreeing about it.

I need a drink.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4605 d3437ac4-8b1b-0410-be28-b75ab53cb89f