Corrected -[UITableView _scrollRectToVisible:atScrollPosition:animated:];  UITableViewScrollPositionNone was not being handled properly.  From the Apple docs for UITableViewScrollPositionNone:
The table view scrolls the row of interest to be fully visible with a minimum of movement. If the row is already fully visible, no scrolling occurs. For example, if the row is above the visible area, the behavior is identical to that specified by UITableViewScrollPositionTop. This is the default.