Implemented -[UINib instantiateWithOwner:options:]
This is much better.  At this point, all of the basic UINib machinery should be in place and functional.  Right now, we *should* be properly loading:

UIButton
UIColor
UIFont
UIView

In order to have the nib-support machinery load a class, what's necessary is to add conformance to the NSCoding protocol -- specifically the -initWithCoder: constructor.  The -encodeWithCoder: method should be implemented in terms of [self doesNotRespondToSelector:_cmd] for view classes, since reading is all that's required.

Practically speaking, you'll need to know which keys are set for each view by IB and what their types are.  No need to guess here.

1) Construct a xib with IB that contains your target class, and preferably little else.
2) Change an option on the view.
3) Compile the xib.
4) Locate the resulting .nib, and convert it to XML using "plutil -convert xml1"
5) Open this with your favorite text editor and look it over.  It should be pretty apparent what the names and types are.

Enjoy.