First pass at support for iOS style "NIBArchive" nibs.
This didn't come easily.  I had to reverse engineer the file format, and stare at hex until I understood what was happening.  There aren't any public reference materials or other open source projects to refer to, so it's more than likely that there are bugs and misconceptions in here.  It *does* load the view-graphs for our app, though, so we're definitely off to a good start.